requirements/common.txt
Django
psycopg2-binary
djangorestframework
django-rest-knox